+++ test/src/utils/random_utils_test.dart
@@ -7,13 +7,12 @@ void main() {
     expect(result, 1);
   });
 
-
-   test('get  random array elements', () {
+  test('get  random array elements', () {
     final result = RandomUtils.arrayElements([1], 3);
     expect(result, [1, 1, 1]);
   });
 
-    test('get  random array elements from empty array or null', () {
+  test('get  random array elements from empty array or null', () {
     final result = RandomUtils.arrayElements(null, 3);
     expect(result, []);
   });